refactor: rename runner-go -> runner
Instead of having per-language runners, we'll have a single service with
a configuration file telling it how to launch specific language
runtimes.
